Major rewrite to allow for use on any IoT dev board.
Primarily rewrote to eliminate the need for an I/O sensor, so one can get this app up and running quickly and without problems, even on a NUC or Gateway device. Also rewrote the way the app delivers data to the browser, eliminating the need for the index.html file. Changed web server configuration so it uses a generic 0.0.0.0 IP address, which accepts requests from any open and active nework interface.